PRSD-945: Moves FileUploaders into service folders
